+++ keyboards/keychron/q1/rev_0100/keymaps/gtg465x/rgb_matrix_user.h
@@ -18,7 +18,7 @@
 
 void rgb_matrix_init_user(void);
 
-void rgb_matrix_set_color_by_keycode(uint8_t led_min, uint8_t led_max, uint8_t layer, bool (*is_keycode)(uint8_t), uint8_t red, uint8_t green, uint8_t blue);
+void rgb_matrix_set_color_by_keycode(uint8_t led_min, uint8_t led_max, uint8_t layer, bool (*is_keycode)(uint16_t), uint8_t red, uint8_t green, uint8_t blue);
 
-bool is_caps_lock_indicator(uint8_t keycode);
-bool is_transparent(uint8_t keycode);
+bool is_caps_lock_indicator(uint16_t keycode);
+bool is_transparent(uint16_t keycode);
